+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.5.4](https://github.com/dldevinc/paper-rq/tree/v0.5.4) - 2022-04-12
 ### Bug Fixes
 - Fix an issue with expired jobs in the queue.
+- Ignore jobs that raise a `DeserializationError`.
 
 ## [0.5.3](https://github.com/dldevinc/paper-rq/tree/v0.5.3) - 2022-03-23
 ### Bug Fixes